requirements.txt
Markdown==2.6.5
Pillow==3.1.1
psycopg2==2.6.1
django-filter==0.12.0
djangorestframework==3.3.2
coverage==4.0.3
Django==1.9.1
